Vastly improve performance of IRQ handling on stopped interfaces with this driver by introducing a flag saying we already stopped the device.
On my Soekris net4801, this took a ping -i 0.001 from spending 80% of
time in interrupt handling to 10% (approx numbers).

This was a particular problem for the net4801 because the tree
interfaces share the same interrupt, but it would be a problem for
any configuration where an unused if_sis interface shares an interrupt
with a busy device.

Other drivers may have similar problems.

Thanks to:	Luigi


git-svn-id: svn+ssh://svn.freebsd.org/base/head@123833 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f