Under a heavy RX load, at least with D-Link DFE-550TX adapters, the driver's RX ring head may fall behind the chip, causing the stuck traffic, disordered packets, etc.  Work around this by adopting the technique of resyncing RX head used in dc(4) and xl(4) drivers, but do it in a slightly different place to reduce the number of resyncs needed.
Also, set the NIC's RX polling period to a more meaningful value,
to stop overloading the PCI bus (this also reduces the number of
resyncs by a factor of 3 or more in a long run; the actual number
is very dependent on a nature of the traffic).

Maintain the statistics counter as the hw.ste_rxsyncs sysctl.

In cooperation with:	Vsevolod Lobko
OK'ed by:		ambrisko
MFC after:		5 days