Make sure that symbols are declared in math.h iff the appropriate namespaces are visible.  Previously, math.h failed to hide some C99-, XSI-, and BSD-specific symbols in certain compilation environments.
The referenced PR has a nice listing of the appropriate conditions for
making symbols visible in math.h.  The only non-stylistic difference
between the patch in the PR and this commit is that I superfluously
test for __BSD_VISIBLE in a few places to be more explicit about which
symbols have historically been part of the FreeBSD environment.

PR:		65939
Submitted by:	Stefan Farfeleder <stefan@fafoe.narf.at>


git-svn-id: svn+ssh://svn.freebsd.org/base/head@128628 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f