Auto generate the device listings in the Hardware Notes based on manual pages, to avoid duplicating work between the Hardware Notes and manual pages.
Each text line in the output from the manual page parser is generated
as a SGML entity, making it possible for translators to only translate
lines which actually contains English text (this neat idea came from
hrs).

To determine which drivers are usable on which architectures the
dev.archlist.txt file is used to explicitly list this.  It it an
imperfect solution, but the best I have been able to come up with for
now.

This commit converts most of the devices which has a section 4 manual
page with a HARDWARE section to be the new auto generated format.

Reviewed in principle by:	ru, hrs, trhodes
Good ideas for improvements:	hrs
No objections:			-doc, re


git-svn-id: svn+ssh://svn.freebsd.org/base/head@133094 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f