Remove an XXXRW indicating atomic operations might be used as a substitute for a global mutex protecting the socket count and generation number.
The observation that soreceive_rcvoob() can't return an mbuf
chain is a property, not a bug, so remove the XXXRW.

In sorflush, s/existing/previous/ for code when describing prior
behavior.

For SO_LINGER socket option retrieval, remove an XXXRW about why
we hold the mutex: this is correct and not dubious.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@139216 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f