Fix support for Windows drivers that support both PCI and PCMCIA devices at the same time.
Fix if_ndis_pccard.c so that it sets sc->ndis_dobj and sc->ndis_regvals.

Correct IMPORT_SFUNC() macros for the READ_PORT_BUFFER_xxx() routines,
which take 3 arguments, not 2.

This fixes it so that the Windows driver for my Cisco Aironet 340 PCMCIA
card works again. (Yes, I know the an(4) driver supports this card natively,
but it's the only PCMCIA device I have with a Windows XP driver.)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@146015 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f