Add the full module path name to the kld_file_stat structure for kldstat(2).
This allows libdtrace to determine the exact file from which
a kernel module was loaded without having to guess.

The kldstat(2) API is versioned with the size of the
kld_file_stat structure, so this change creates version 2.

Add the pathname to the verbose output of kldstat(8) too.

MFC: 3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@172862 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f