Add another new sysctl in support of the forthcoming procstat(1) to support its -k argument:
kern.proc.kstack - dump the kernel stack of a process, if debugging
  is permitted.

This sysctl is present if either "options DDB" or "options STACK" is
compiled into the kernel.  Having support for tracing the kernel
stacks of processes from user space makes it much easier to debug
(or understand) specific wmesg's while avoiding the need to enter
DDB in order to determine the path by which a process came to be
blocked on a particular wait channel or lock.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@174197 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f