Add support for the Oxford OX16PCI958-based card.
Note, that the patch provided with this card for the Linux states that
the card uses DEFAULT_RCLK * 2, while in fact it is '* 10'.  So probably
we should also use the subdevice/subvendord here. For now just ignore
that fact.

PR:		kern/129665
Submitted by:	bsam
Obtained from:	united efforst with bsam


git-svn-id: svn+ssh://svn.freebsd.org/base/head@186520 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f