Add missing POSIX 1003.1-2008 open(2) flag; O_TTY_INIT.
On FreeBSD, this is the default behaviour. According to the spec, we may
give this flag a value of zero, but I'd rather not do this. If we define
it to a non-zero value, we can always change default behaviour without
changing the ABI. This is very unlikely to happen, though.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@189143 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f