Add ipx_ifaddr locking to ipx_control(), which should close most remaining potential races in ifconfig's management of IPX addresses.
This is largely accomplished by dropping a global write lock for the
IPX address list over the body of in_control(), although there are
some places we bump the refcount on an ifaddr of interest while
calling out to the routing code or link layer code, which might
require revisiting.

Annotate one as a potential race if two simultaneous delete ioctls
are issued for the same IPX addresses at once.

MFC after:	3 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@194610 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f