On amd64, change clang's default triple to 'x86_64-unknown-freebsd9.0', similar to what we do for binutils.  When clang's default triple starts with 'amd64-', it does not pass a proper -target-cpu option to its first stage.
This can lead to problems, for example when structs are memcpy'd, and
clang erroneously assumes they are 16-byte aligned.  It will then use
the 'movaps' SSE instruction to implement the copy, which results in a
bus error if the struct is really 8-byte aligned.

I encountered this issue when gcc's /usr/libexec/cc1 started crashing
with SIGBUS, after rebuilding world with clang ToT, but it also affects
the version of clang that we have in the tree.  We were just lucky until
now, apparently. :)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@222183 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f