The AR5212 setup path (also used by the AR5416 code) configures a local variable with a beacon interval of 100 TU. This never gets modified if the beacon interval configuration changes.
This may have been correct in earlier times, but with the advent of
staggered beacons (which default to 1 / ATH_BCBUF beacon interval, so
25 TU here) this value is incorrect.

It is used to configure the default CABQ readytime. So here, the cabq
was being configured to be much greater than the target beacon timer
(TBTT.)

The driver should be configuring a cabq readytime value rather then
leaving it to the HAL to choose sensible defaults. This should be
done in the future - I'm simply trying to ensure sensible defaults
are chosen.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@225819 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f