Remove the now unused skpc() function.
It was only used by ufs and ext2 and I have really strong doubts that
there are other pieces of code that also use this function. If it turns
out that external drivers use this code as well, I'd be happy to migrate
or revert.

Bump __FreeBSD_version while there.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@229204 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f