Invalidate caches immediately upon entry to init_secondary().  Also set the Bufferable bit in the PDE entries of the secondary processor startup pagetables.
The caches really need to be invalidated even earlier than this, but this
is a big step in the right direction.  The invalidate needs to happen
before the MMU is enabled, which means it has to be called from asm code
that's running with physical addressing.  Fixing that will be handled in
a future change.