sys/capsicum.h: Remove not very useful visibility guards.
Since <sys/capsicum.h> does not comply to any standards, standards-compliant
applications do not include it and it is not useful to restrict what is
exposed depending on the requested standard.

In addition, the use of types like u_int makes the header fail in strict
standards compliance modes anyway.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@264683 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f