Flush and invalidate caches on each CPU as part of handling IPI_STOP.
Flushing the caches is required before doing a panic dump, but ARM
doesn't provide a flavor of flush that gets broadcast to other cores.
However, all cores except one are stopped before doing a dump, so this
works around the lack of a global flush/invalidate by doing it locally
on each CPU as part of stopping.

Discussed with:	cognet@


git-svn-id: svn+ssh://svn.freebsd.org/base/head@265024 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f