Document some user-settable make variables in ports.7
This is not a comprehensive list, as the variables themselves are spread
out over multiple files, but it is a start.

Add a section to make.conf noting that variables may be set there that
affect ports builds, but refer to ports.7 and elsewhere for the actual
listing; any listing in make.conf.5 would likely become out of date
fairly quickly.

PR:		docs/189199
Reviewed by:	bdrewery (previous version)
Approved by:	hrs (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@266416 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f