Make igb and ixgbe check tunables at probe time.
This allows one to make a kernel module to tune the
number of queues before the driver loads.

This is needed so that a module at SI_SUB_CPU can set
tunables for these drivers to take.  Otherwise getenv
is called too early by the TUNABLE macros.

Reviewed by: smh
Phabric: https://reviews.freebsd.org/D1149


git-svn-id: svn+ssh://svn.freebsd.org/base/head@275136 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f