Partial MFC r281838:
For igb(4), when we are doing multiqueue, we are all setup to have full 32bit
RSS hash from the card. Expose that so others like lagg(4) can use that and
avoid hashing the traffic by themselves. Setting hashtype as OPAQUE because
FreeBSD 10 doesn't have RSS support.

Sponsored by:	Limelight Networks