In case of an output error, continue with the next net, don't try to continue sending on the same net.
This fixes a bug where an invalid mbuf chain was constructed, if a
full size frame of control chunks should be sent and there is a
output error.

Based on a discussion with rrs@, change move to the next net. This fixes
the bug and improves the behaviour.

Thanks to Irene Ruengeler for spending a lot of time in narrowing this
problem down.
MFC after: 3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@284326 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f