Add a hack-around to this fatal taskqueue running whilst the NIC is detaching.
This mostly fixes a panic - the reset path shouldn't run whilst
the NIC is being torn down.

It's not locked, so it's "mostly" ok, but most of the rest of
the driver doesn't read sc->invalid with sensible locking. Grr.

The real solution is to cleanly tear down taskqueues in the detach/suspend
phase, but ..


git-svn-id: svn+ssh://svn.freebsd.org/base/head@286343 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f